+++ src/main/kotlin/me/mattco/serenityos/gml/GML.kt
@@ -1,0 +1,26 @@
+package me.mattco.serenityos.gml
+
+import com.intellij.extapi.psi.PsiFileBase
+import com.intellij.icons.AllIcons
+import com.intellij.lang.Language
+import com.intellij.openapi.fileTypes.LanguageFileType
+import com.intellij.openapi.util.IconLoader
+import com.intellij.psi.FileViewProvider
+
+object GMLLanguage : Language("SerenityOS GML") {
+    val FILE_ICON = IconLoader.getIcon("/META-INF/gml.png", GMLLanguage::class.java)
+}
+
+object GMLFileType : LanguageFileType(GMLLanguage) {
+    override fun getName() = "SerenityOS GML"
+
+    override fun getDescription() = "The SerenityOS GML file type"
+
+    override fun getDefaultExtension() = "gml"
+
+    override fun getIcon() = GMLLanguage.FILE_ICON
+}
+
+class GMLFile(viewProvider: FileViewProvider) : PsiFileBase(viewProvider, GMLLanguage) {
+    override fun getFileType() = GMLFileType
+}