+++ src/main/kotlin/me/mattco/serenityos/gml/GMLFileStub.kt
@@ -1,0 +1,11 @@
+package me.mattco.serenityos.gml
+
+import com.intellij.psi.stubs.PsiFileStubImpl
+import com.intellij.psi.tree.IStubFileElementType
+
+class GMLFileStub(file: GMLFile?) : PsiFileStubImpl<GMLFile>(file) {
+    object Type : IStubFileElementType<GMLFileStub>("GMLFile", GMLLanguage) {
+        // Should be incremented when lexer, parser, or stub tree changes
+        override fun getStubVersion() = 1
+    }
+}