+++ src/main/kotlin/me/mattco/serenityos/gml/GMLLineMarkerProvider.kt
@@ -1,0 +1,26 @@
+package me.mattco.serenityos.gml
+
+import com.intellij.codeInsight.daemon.LineMarkerInfo
+import com.intellij.codeInsight.daemon.LineMarkerProvider
+import com.intellij.codeInsight.navigation.NavigationGutterIconBuilder
+import com.intellij.icons.AllIcons
+import com.intellij.psi.PsiElement
+import com.jetbrains.cidr.lang.psi.OCStructLike
+import me.mattco.serenityos.gml.psi.api.GMLComponent
+
+class GMLLineMarkerProvider : LineMarkerProvider {
+    override fun collectSlowLineMarkers(
+        elements: MutableList<out PsiElement>,
+        result: MutableCollection<in LineMarkerInfo<*>>
+    ) {
+        elements.filterIsInstance<GMLComponent>().forEach {
+            NavigationGutterIconBuilder.create(AllIcons.Nodes.ModelClass)
+                .setTarget(it.reference?.resolve() as? OCStructLike ?: return@forEach)
+                .setTooltipText("Jump to C++ Widget")
+                .createLineMarkerInfo(it)
+                .let(result::add)
+        }
+    }
+
+    override fun getLineMarkerInfo(element: PsiElement): LineMarkerInfo<*>? = null
+}