+++ src/main/kotlin/me/mattco/serenityos/gml/psi/GMLNameIdentifierOwner.kt
@@ -1,0 +1,6 @@
+package me.mattco.serenityos.gml.psi
+
+import com.intellij.psi.NavigatablePsiElement
+import com.intellij.psi.PsiNameIdentifierOwner
+
+interface GMLNameIdentifierOwner : GMLPsiElement, PsiNameIdentifierOwner, NavigatablePsiElement