+++ src/main/kotlin/me/mattco/serenityos/gml/psi/GMLPsiElement.kt
@@ -1,0 +1,5 @@
+package me.mattco.serenityos.gml.psi
+
+import com.intellij.psi.PsiElement
+
+interface GMLPsiElement : PsiElement